+++ extension/src/Settings.ts
@@ -32,6 +32,9 @@ export enum Setting {
     DocsIgnorePaths,
     DocsRootPath,
     CreateTocFilesForDocs,
+    IncludeTablesAndFieldsInDocs,
+    CreateInfoFileForDocs,
+    CreateUidForDocs,
     XliffCSVExportPath,
     XliffCSVImportTargetState
 }
@@ -75,6 +78,9 @@ export class Settings {
         this.SettingCollection[Setting.RemoveObjectNamePrefixFromDocs] = this.config.get('RemoveObjectNamePrefixFromDocs');
         this.SettingCollection[Setting.DocsIgnorePaths] = this.config.get('DocsIgnorePaths');
         this.SettingCollection[Setting.CreateTocFilesForDocs] = this.config.get('CreateTocFilesForDocs');
+        this.SettingCollection[Setting.IncludeTablesAndFieldsInDocs] = this.config.get('IncludeTablesAndFieldsInDocs');
+        this.SettingCollection[Setting.CreateInfoFileForDocs] = this.config.get('CreateInfoFileForDocs');
+        this.SettingCollection[Setting.CreateUidForDocs] = this.config.get('CreateUidForDocs');
         this.SettingCollection[Setting.XliffCSVExportPath] = this.config.get('Xliff CSV Export Path');
         this.SettingCollection[Setting.XliffCSVImportTargetState] = this.config.get('Xliff CSV Import Target State');
     }