+++ extension/src/test/ALXmlComment.test.ts
@@ -5,19 +5,21 @@ import { ALProcedure } from '../ALObject/ALProcedure';
 import * as ALObjectTestLibrary from './ALObjectTestLibrary';
 suite("Classes.AL Functions Tests", function () {
     test("XML Comments No Line Breaks formatting", function () {
-        assert.equal(ALXmlComment.formatMarkDown('asfd <para>bold 1</para> sadf <para>bold 2</para> asdf', true), `asfd   bold 1   sadf   bold 2   asdf`, 'Unexpected paragraph');
-        assert.equal(ALXmlComment.formatMarkDown('asfd <b>bold 1</b> sadf <b>bold 2</b> asdf', true), 'asfd **bold 1** sadf **bold 2** asdf', 'Unexpected bold');
-        assert.equal(ALXmlComment.formatMarkDown('asfd <i>italic 1</i> sadf <i>italic 2</i> asdf', true), 'asfd *italic 1* sadf *italic 2* asdf', 'Unexpected italic');
-        assert.equal(ALXmlComment.formatMarkDown('asfd <c>code 1</c> sadf <c>code 2</c> asdf', true), 'asfd \`code 1\` sadf \`code 2\` asdf', 'Unexpected code');
-        assert.equal(ALXmlComment.formatMarkDown(`asfd <code>code block 1
+        assert.equal(ALXmlComment.formatMarkDown({ text: 'asfd <para>bold 1</para> sadf <para>bold 2</para> asdf', inTableCell: true }), `asfd   bold 1   sadf   bold 2   asdf`, 'Unexpected paragraph');
+        assert.equal(ALXmlComment.formatMarkDown({ text: 'asfd <b>bold 1</b> sadf <b>bold 2</b> asdf', inTableCell: true }), 'asfd **bold 1** sadf **bold 2** asdf', 'Unexpected bold');
+        assert.equal(ALXmlComment.formatMarkDown({ text: 'asfd <i>italic 1</i> sadf <i>italic 2</i> asdf', inTableCell: true }), 'asfd *italic 1* sadf *italic 2* asdf', 'Unexpected italic');
+        assert.equal(ALXmlComment.formatMarkDown({ text: 'asfd <c>code 1</c> sadf <c>code 2</c> asdf', inTableCell: true }), 'asfd \`code 1\` sadf \`code 2\` asdf', 'Unexpected code');
+        assert.equal(ALXmlComment.formatMarkDown({
+            text: `asfd <code>code block 1
 asdf afd</code>
 sadf
 <code>code block 2</code>
-asdf`, true), `asfd \`code block 1`, 'Unexpected code block');
+asdf`, inTableCell: true
+        }), `asfd \`code block 1`, 'Unexpected code block');
     });
 
     test("XML Comments formatting", function () {
-        assert.equal(ALXmlComment.formatMarkDown('asfd <para>bold 1</para> sadf <para>bold 2</para> asdf'), `asfd 
+        assert.equal(ALXmlComment.formatMarkDown({ text: 'asfd <para>bold 1</para> sadf <para>bold 2</para> asdf' }), `asfd 
 
 bold 1
 
@@ -26,15 +28,16 @@ bold 1
 bold 2
 
  asdf`, 'Unexpected paragraph');
-        assert.equal(ALXmlComment.formatMarkDown('asfd <b>bold 1</b> sadf <b>bold 2</b> asdf'), 'asfd **bold 1** sadf **bold 2** asdf', 'Unexpected bold');
-        assert.equal(ALXmlComment.formatMarkDown('asfd <i>italic 1</i> sadf <i>italic 2</i> asdf'), 'asfd *italic 1* sadf *italic 2* asdf', 'Unexpected italic');
-        assert.equal(ALXmlComment.formatMarkDown('asfd <c>code 1</c> sadf <c>code 2</c> asdf'), 'asfd \`code 1\` sadf \`code 2\` asdf', 'Unexpected code');
-        assert.equal(ALXmlComment.formatMarkDown(`asfd 
+        assert.equal(ALXmlComment.formatMarkDown({ text: 'asfd <b>bold 1</b> sadf <b>bold 2</b> asdf' }), 'asfd **bold 1** sadf **bold 2** asdf', 'Unexpected bold');
+        assert.equal(ALXmlComment.formatMarkDown({ text: 'asfd <i>italic 1</i> sadf <i>italic 2</i> asdf' }), 'asfd *italic 1* sadf *italic 2* asdf', 'Unexpected italic');
+        assert.equal(ALXmlComment.formatMarkDown({ text: 'asfd <c>code 1</c> sadf <c>code 2</c> asdf' }), 'asfd \`code 1\` sadf \`code 2\` asdf', 'Unexpected code');
+        assert.equal(ALXmlComment.formatMarkDown({
+            text: `asfd 
 <code>code block 1
 asdf afd</code>
 sadf
 <code>code block 2</code>
-asdf`), `asfd 
+asdf` }), `asfd 
 
 \`\`\`javascript
 code block 1